mgr/cephadm: Add NFSService
